pyproject.toml
[project]
name = "minimart"
version = "1.0.0"
description = "Multi-tenant shopping list application with real-time synchronization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.128.4",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.46",
    "asyncpg>=0.30.0",
    "alembic>=1.15.1",
    
    # Redis
    "redis>=5.2.1",
    
    # Authentication
    "python-jose[cryptography]>=3.4.0",
    "argon2-cffi>=25.1.0",
    
    # Email
    "aiosmtplib>=3.0.3",
    "email-validator>=2.2.0",
    
    # Validation & Settings
    "pydantic>=2.12.5",
    "pydantic-settings>=2.8.1",
    
    # Utilities
    "python-dotenv>=1.1.0",
    "uuid6>=2024.2.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "httpx>=0.28.1",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "httpx>=0.28.1",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
